Create report conductor with police button
Needs repairing the calculation of chances in client (danger-indicator.component).

+++ server/src/main/migrations/20180211151800-create-route-point.ts
@@ -18,7 +18,7 @@ export = new NewTableMigration("RoutePoints", (types: DataTypes): DefineAttribut
          model: "Routes",
          key: "id",
       },
-      type: types.INTEGER,
+      type: types.BIGINT,
    },
 
    stationId: {
@@ -27,6 +27,6 @@ export = new NewTableMigration("RoutePoints", (types: DataTypes): DefineAttribut
          model: "Stations",
          key: "id",
       },
-      type: types.INTEGER,
+      type: types.BIGINT,
    },
 }));
